+++ test/pylib/scylla_cluster.py
@@ -25,7 +25,7 @@ from io import BufferedWriter
 from test.pylib.host_registry import Host, HostRegistry
 from test.pylib.pool import Pool
 from test.pylib.rest_client import ScyllaRESTAPIClient, HTTPError
-from test.pylib.util import LogPrefixAdapter
+from test.pylib.util import LogPrefixAdapter, read_last_line
 from test.pylib.internal_types import ServerNum, IPAddress, HostID, ServerInfo
 import aiohttp
 import aiohttp.web
@@ -401,29 +401,6 @@ class ScyllaServer:
         self.start_time = time.time()
         sleep_interval = 0.1
         cql_up_state = CqlUpState.NOT_CONNECTED
-
-        def read_last_line(file_path: pathlib.Path):
-            block_size = 4 * 1024
-            file_size = os.stat(file_path).st_size
-            pos = file_size
-            blocks = []
-            linesep = os.linesep.encode()
-            with file_path.open('rb') as f:
-                linesep_index = -1
-                while pos > 0 and linesep_index == -1:
-                    next_pos = max(pos - block_size, 0)
-                    f.seek(next_pos, os.SEEK_SET)
-                    block = f.read(pos - next_pos)
-                    # ignore the last empty line if any
-                    if pos == file_size and block.endswith(linesep):
-                        block = block[:-len(linesep)]
-                    linesep_index = block.rfind(linesep)
-                    blocks.append(block)
-                    pos = next_pos
-            if linesep_index != -1:
-                blocks[-1] = block[linesep_index + len(linesep):]
-            blocks.reverse()
-            return b''.join(blocks).decode()
 
         def report_error(message: str):
             message += f", server_id {self.server_id}, IP {self.ip_addr}, workdir {self.workdir.name}"
